refactor homepage into website app
