+++ ophasebase/templates/ophasebase/base.html
@@ -15,7 +15,7 @@
         <nav class="navbar navbar-default" role="navigation">
             <div class="container-fluid">
                 <div class="navbar-header">
-                    <a class="navbar-brand" href="{% url 'ophasebase:welcome' %}">OphaseHQ</a>
+                    <a class="navbar-brand" href="{% url 'website:homepage' %}">OphaseHQ</a>
                 </div>
                 <div class="navbar-right">
                     {% if user.is_authenticated %}
