+++ ophasebase/urls.py
@@ -1,8 +1,3 @@
-from django.conf.urls import url
-
-from ophasebase import views
-
 app_name = 'ophasebase'
-urlpatterns = [
-    url(r'^$', views.WelcomeView.as_view(), name='welcome'),
-]
+
+urlpatterns = []
